Launch PA with different args
-F to use system.pa (since in system mode) instead of default.pa
(which PA docs say is for user mode)

Note: Merely declaing PULSE_SCRIPT to point at system.pa in
snapcraft.yaml does not seem to work, so we need the flag.

Remove -p to specify module pluging path. this is declared already
through PULSE_DLPATH var in snapcraft.yaml so unless it is later
found necessary, drop it.