comment out second hw device for usb
This system.pa adds the hw 0,0 alsa sink, which is the output jack
on pi3. Comment out add of hw 1,0, which adds alsa sink for usb
speaker. We also have the the bluetooth modules loaded, and many of
the conditional module loads (if the module SO file exists)
commented OUT, for example udev, etc.